refactor(analytics): split 1922-line engine into domain mixins
AnalyticsEngine was a 1922-line god-class with 80 methods spanning 10
distinct domains. Split into a core/analytics/ package composed by MRO:

  base.py           -- shared state, prefs, property accessors, FY helper
  classification.py -- category predicates (_is_salary/_is_investment/etc.)
  summaries.py      -- monthly/daily summary persistence
  trends.py         -- category trends + transfer flows
  merchants.py      -- merchant extraction + recurring detection helpers
  recurring.py      -- recurring-pattern detection
  net_worth.py      -- snapshot + investment holdings + bucketing
  fy_summaries.py   -- fiscal-year rollups with YoY changes
  anomalies.py      -- anomaly detection + budget tracking
  engine.py         -- composed AnalyticsEngine(*mixins) + run_full_analytics
                      orchestrator + _log_audit

core/analytics_engine.py is now a 13-line facade re-exporting AnalyticsEngine
from the package, so all three production callers (sync_engine.py,
analytics_v2.py, core/__init__.py) keep working unchanged.

File sizes: 1922 -> 11 files (61-300 lines each). Every mixin inherits
from AnalyticsEngineBase for typing only; Python MRO collapses them at
runtime so __init__ runs once.

Also deletes REFACTOR_PLAN.md -- every item in it is now done.

Verified: mypy (strict, 101 files), ruff, all 70 tests pass.
